Add files via upload
This project is a website called "JoJo's Stand Generator" designed for fans of the popular anime and manga series "JoJo's Bizarre Adventure." The website allows users to generate their unique "stands," which are supernatural entities with special abilities featured in the series.

The main features of the website include a homepage that welcomes users and provides directions on how to generate their stands. Additionally, there is a login page where users can log in to access additional features, and when logged in, they are greeted personally on the homepage.

The website is built using a combination of HTML, Python, SQLite and minimal CSS. HTML is used for creating the structure and layout of the web pages, while Python handles the server-side functionality and database interactions using the Flask framework. The SQLite database stores user information, such as usernames and passwords, for the login functionality.